+++ tests/features/fixture_utils.py
@@ -1,7 +1,3 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-from __future__ import print_function
-
 import os
 import codecs
 
@@ -22,7 +18,7 @@ def read_fixture_files():
     """Read all files inside fixture_data directory."""
     current_dir = os.path.dirname(__file__)
     fixture_dir = os.path.join(current_dir, "fixture_data/")
-    print ("reading fixture data: {}".format(fixture_dir))
+    print("reading fixture data: {}".format(fixture_dir))
     fixture_dict = {}
     for filename in os.listdir(fixture_dir):
         if filename not in [".", ".."]: