+++ tests/features/steps/basic_commands.py
@@ -1,10 +1,8 @@
-# -*- coding: utf-8
 """
 Steps for behavioral style tests are defined in this module.
 Each step is defined by the string decorating it.
 This string is used to call the step in "*.feature" file.
 """
-from __future__ import unicode_literals, print_function
 
 import pexpect
 import subprocess
@@ -80,6 +78,19 @@ def step_send_help(context):
     Send \? to see help.
     """
     context.cli.sendline("\?")
+
+
+@when("we send partial select command")
+def step_send_partial_select_command(context):
+    """
+    Send `SELECT a` to see completion.
+    """
+    context.cli.sendline("SELECT a")
+
+
+@then("we see error message")
+def step_see_error_message(context):
+    wrappers.expect_exact(context, 'column "a" does not exist', timeout=2)
 
 
 @when("we send source command")