+++ tests/test_smart_completion_multiple_schemata.py
@@ -1,5 +1,3 @@
-from __future__ import unicode_literals, print_function
-
 import itertools
 from metadata import (
     MetaData,
@@ -229,7 +227,7 @@ def test_suggested_column_names_in_function(completer):
         completer, "SELECT MAX( from custom.products", len("SELECT MAX(")
     )
     assert completions_to_set(result) == completions_to_set(
-        testdata.columns("products", "custom")
+        testdata.columns_functions_and_keywords("products", "custom")
     )